Arreglo de los modulos de css
