+++ src/app/globals.css
@@ -48,176 +48,3 @@ body {
   }
 }
 
-/* form styles */
-
-/* Contenedor principal que centra el formulario en la página */
-.login-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh; /* Ocupa al menos toda la altura de la pantalla */
-  padding: 1rem; /* Espacio para que el formulario no toque los bordes en móviles */
-}
-
-/* El "card" o tarjeta del formulario */
-.login-form {
-  display: flex;
-  flex-direction: column; /* Apila los elementos verticalmente */
-  gap: 1.25rem; /* Espacio consistente entre todos los elementos hijos */
-  width: 100%;
-  max-width: 400px; /* Evita que el formulario sea demasiado ancho en pantallas grandes */
-  padding: 2rem; /* Espaciado interno generoso */
-  border: 1px solid; /* El color lo tomará de tu tema */
-  border-radius: 0.5rem; /* Bordes redondeados */
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05); /* Sombra sutil opcional */
-}
-
-/* Título del formulario */
-.login-title {
-  font-size: 1.75rem;
-  font-weight: 600; /* Semi-bold */
-  text-align: center;
-  margin: 0; /* Quitamos el margen por defecto para que 'gap' controle el espacio */
-}
-
-/* Estilos para los campos de texto y contraseña */
-.login-form input[type="text"],
-.login-form input[type="password"] {
-  width: 100%;
-  padding: 0.75rem 1rem;
-  font-size: 1rem;
-  border: 1px solid; /* El color vendrá de tu tema */
-  border-radius: 0.375rem; /* Bordes redondeados más sutiles */
-  box-sizing: border-box; /* Asegura que el padding no afecte el ancho total */
-}
-
-/* Estilos para el botón principal */
-.btn-primary {
-  padding: 0.75rem 1.5rem;
-  font-size: 1rem;
-  font-weight: 500;
-  text-align: center;
-  cursor: pointer;
-  border: 1px solid transparent; /* Borde transparente por si quieres añadirlo en hover */
-  border-radius: 0.375rem;
-  transition: opacity 0.2s ease-in-out, transform 0.2s ease-in-out;
-}
-
-/* Efecto al pasar el mouse (sin usar colores) */
-.btn-primary:hover {
-  opacity: 0.9;
-  transform: scale(1.02); /* Un ligero efecto de crecimiento */
-}
-
-/* Texto de registro */
-.register-text {
-  text-align: center;
-  font-size: 0.875rem;
-  margin: 0;
-}
-
-/* Estilo del enlace */
-.link {
-  font-weight: 500;
-  text-decoration: underline;
-  text-decoration-offset: 2px;
-}
-
-/* Clase de utilidad para accesibilidad (Screen Reader Only) */
-/* Esto oculta visualmente un elemento pero lo mantiene accesible para lectores de pantalla */
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border-width: 0;
-}
-
-
-/* === ESTILOS PARA PANTALLAS MÁS GRANDES (TABLETS Y DESKTOPS) 💻 === */
-
-/* A partir de 768px, podemos hacer el padding del formulario un poco más grande */
-@media (min-width: 768px) {
-  .login-form {
-    padding: 2.5rem;
-    gap: 1.5rem;
-  }
-
-  .login-title {
-    font-size: 2rem;
-  }
-}
-
-/* Container */
-
-/* === ESTILOS BASE (MOBILE-FIRST) 📱 === */
-
-/* Contenedor principal de la página */
-.container {
-  display: flex;
-  flex-direction: column; /* Apila la sección de bienvenida y el login verticalmente */
-  align-items: center; /* Centra todo horizontalmente */
-  justify-content: center; /* Centra el contenido en el medio de la pantalla */
-  min-height: 100vh; /* Ocupa al menos toda la altura de la pantalla */
-  padding: 2rem; /* Espaciado para que no toque los bordes */
-  gap: 3rem; /* Espacio vertical entre la sección del logo y el formulario de login */
-}
-
-/* Div que contiene la información de la app */
-.container > div:first-child {
-  display: flex;
-  flex-direction: column;
-  align-items: center; /* Centra el logo y el texto */
-  text-align: center;
-}
-
-/* Estilo para el logo */
-.logo-image {
-  /* Next.js Image maneja bien el tamaño, pero podemos asegurar que no se encoja */
-  flex-shrink: 0;
-}
-
-/* Título principal de la app */
-.app-title {
-  font-size: 2.5rem;
-  font-weight: 700;
-  margin-top: 1rem;
-  margin-bottom: 0.5rem;
-}
-
-/* Descripción de la app */
-.app-description {
-  font-size: 1.125rem; /* 18px */
-  max-width: 350px; /* Evita que el texto sea demasiado largo en una línea */
-  opacity: 0.8; /* Le da un tono más suave que el título */
-  margin: 0;
-}
-
-/* === ESTILOS PARA PANTALLAS GRANDES (DESKTOPS) 💻 === */
-
-/* Usamos 992px como punto de quiebre para un layout de escritorio */
-@media (min-width: 992px) {
-  .container {
-    flex-direction: row; /* Cambia a un layout horizontal (lado a lado) */
-    justify-content: space-evenly; /* Distribuye el espacio entre los dos elementos */
-    gap: 5rem; /* Aumenta el espacio entre la info y el login */
-  }
-
-  /* Alineamos el texto de la sección de bienvenida a la izquierda */
-  .container > div:first-child {
-    align-items: flex-start;
-    text-align: left;
-  }
-
-  .app-title {
-    font-size: 3.5rem; /* Hacemos el título más grande en escritorio */
-  }
-
-  .app-description {
-     max-width: 450px; /* Permitimos que la descripción sea un poco más ancha */
-  }
-}
