+++ src/app/login/Login.module.css
@@ -1,0 +1,69 @@
+/* Container */
+
+/* === ESTILOS BASE (MOBILE-FIRST) 📱 === */
+
+/* Contenedor principal de la página */
+.container {
+  display: flex;
+  flex-direction: column; /* Apila la sección de bienvenida y el login verticalmente */
+  align-items: center; /* Centra todo horizontalmente */
+  justify-content: center; /* Centra el contenido en el medio de la pantalla */
+  min-height: 100vh; /* Ocupa al menos toda la altura de la pantalla */
+  padding: 2rem; /* Espaciado para que no toque los bordes */
+  gap: 3rem; /* Espacio vertical entre la sección del logo y el formulario de login */
+}
+
+/* Div que contiene la información de la app */
+.container > div:first-child {
+  display: flex;
+  flex-direction: column;
+  align-items: center; /* Centra el logo y el texto */
+  text-align: center;
+}
+
+/* Estilo para el logo */
+.logo_image {
+  /* Next.js Image maneja bien el tamaño, pero podemos asegurar que no se encoja */
+  flex-shrink: 0;
+}
+
+/* Título principal de la app */
+.app_title {
+  font-size: 2.5rem;
+  font-weight: 700;
+  margin-top: 1rem;
+  margin-bottom: 0.5rem;
+}
+
+/* Descripción de la app */
+.app_description {
+  font-size: 1.125rem; /* 18px */
+  max-width: 350px; /* Evita que el texto sea demasiado largo en una línea */
+  opacity: 0.8; /* Le da un tono más suave que el título */
+  margin: 0;
+}
+
+/* === ESTILOS PARA PANTALLAS GRANDES (DESKTOPS) 💻 === */
+
+/* Usamos 992px como punto de quiebre para un layout de escritorio */
+@media (min-width: 992px) {
+  .container {
+    flex-direction: row; /* Cambia a un layout horizontal (lado a lado) */
+    justify-content: space-evenly; /* Distribuye el espacio entre los dos elementos */
+    gap: 5rem; /* Aumenta el espacio entre la info y el login */
+  }
+
+  /* Alineamos el texto de la sección de bienvenida a la izquierda */
+  .container > div:first-child {
+    align-items: flex-start;
+    text-align: left;
+  }
+
+  .app_title {
+    font-size: 3.5rem; /* Hacemos el título más grande en escritorio */
+  }
+
+  .app_description {
+     max-width: 450px; /* Permitimos que la descripción sea un poco más ancha */
+  }
+}
