+++ src/app/login/page.tsx
@@ -1,24 +1,21 @@
-import Cards from "@/components/conteiners/cards";
-import Login from "@/components/forms/login";
+import Login from "@/components/forms/login/login";
 import Image from "next/image";
+import styles from "./Login.module.css"
 
 export default function Home() {
-  const login = () => {
-    return (<Login />);
-  }
 
   return (
-    <div className="container">
+    <div className={styles.container}>
         <div>
             <Image
             src="/logo.png"
             alt="Logo"
             width={150}
             height={150}
-            className="logo-image"
+            className={styles.logo_image}
             />
-            <h1 className="app-title">Bankpry</h1>
-            <p className="app-description">Tu banco en línea seguro y confiable</p>
+            <h1 className={styles.app_title}>Bankpry</h1>
+            <p className={styles.app_description}>Tu banco en línea seguro y confiable</p>
         </div>
       <Login />
     </div>
