+++ src/app/page.tsx
@@ -1,15 +1,9 @@
-import Cards from "@/components/conteiners/cards";
-import Login from "@/components/forms/login";
+
 
 export default function Home() {
-  const login = () => {
-    return (<Login />);
-  }
 
   return (
     <div>
-      Hola
-      <Cards>{login()}</Cards>
     </div>
   );
 }
