+++ src/components/forms/login/Login.module.css
@@ -1,0 +1,60 @@
+/* Login.module.css */
+
+.login_container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    /* min-height: 100vh; */
+    padding: 1rem;
+    border-color: var(--font-color);
+    border-width: 1px;
+}
+
+.login_form {
+  display: flex;
+  flex-direction: column;
+  gap: 1.25rem;
+  width: 100%;
+  max-width: 400px;
+  padding: 2rem;
+  border: 1px solid rgba(0, 0, 0, 0.1);
+  border-radius: 0.5rem;
+}
+
+.login_title {
+  font-size: 1.75rem;
+  font-weight: 600;
+  text-align: center;
+  margin: 0;
+}
+
+.sr_only {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  margin: -1px;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  white-space: nowrap;
+  border-width: 0;
+}
+
+.btn_primary {
+  padding: 0.75rem 1.5rem;
+  font-size: 1rem;
+  cursor: pointer;
+  border: 1px solid transparent;
+  border-radius: 0.375rem;
+}
+
+.register_text {
+  text-align: center;
+  font-size: 0.875rem;
+  margin: 0;
+}
+
+.link {
+  font-weight: 500;
+  text-decoration: underline;
+}
